third_party/WebKit/LayoutTests/css3/supports.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/js-test.js"></script>
<style>
    .test {
        background-color: blue;
    }

    @supports (display: none) {
        #t0 { background-color: green; }
    }

    @supports (display: deadbeef) {
        #t1 { background-color: red; }
    }

    /* Negation */
    @supports not (display: deadbeef) {
        #t2 { background-color: green; }
    }

    @supports not (display: none) {
        #t3 { background-color: red; }
    }

    @supports not (not (display: none)) {
        #t4 { background-color: green; }
    }

    @supports not (not (not (display: none))) {
        #t5 { background-color: red; }
    }

    /* Conjunction */
    @supports (display: none) and (display: block) {
        #t6 { background-color: green; }
    }

    @supports (display: none) and (display: block) and (display: inline) {
        #t7 { background-color: green; }
    }

    @supports (display: none) and (display: block) and (display: deadbeef) and (display: inline) {
        #t8 { background-color: red; }
    }

    /* Disjunction */
    @supports (display: none) or (display: inline) {
        #t9 { background-color: green; }
    }

    @supports (display: none) or (display: block) or (display: inline) {
        #t10 { background-color: green; }
    }

    @supports (display: none) or (display: deadbeef) or (display: inline) {
        #t11 { background-color: green; }
    }

    @supports (display: ohhai) or (display: deadbeef) or (display: rainbows) {
        #t12 { background-color: red; }
    }

    /* Bad syntax. Can't mix operators without a layer of parentheses. */
    @supports (display: none) and (display: block) or (display: inline) {
        #t13 { background-color: red; }
    }

    @supports not (display: deadbeef) and (display: block) {
        #t14 { background-color: red; }
    }

    /* Mix 'n match */
    @supports (not (border: 1px 1px 1px 1px 1px solid #000)) and (display: block) {
        #t15 { background-color: green; }
    }

    @supports (display: block !important) and ((display: inline) or (display: deadbeef)){
        #t16 { background-color: green; }
    }

    @supports not ((not (display: block)) or ((display: none) and (deadbeef: 1px))) {
        #t17 { background-color: green; }
    }

    /* Whitespace/Syntax */
    @supports not( display: deadbeef) {
        #t22 { background-color: red; }
    }

    @supports (display: none)and (   -webkit-transition: all 1s ) {
        #t23 { background-color: red; }
    }

    @supports (display: none)or(-webkit-transition: all 1s) {
        #t24 { background-color: red; }
    }

    @supports (display: none) or(-webkit-transition: all 1s    ) {
        #t25 { background-color: red; }
    }

    @supports (((((((display: none))))))) {
        #t26 { background-color: green; }
    }

    @supports(((((((display: none))))))) {
        #t27 { background-color: green; }
    }

    @supports (!important) {
        #t28 { background-color: red; }
    }

    @supports not not not not (display: none) {
        #t29 { background-color: red; }
    }

    /* Functions */

    @supports (top: -webkit-calc(80% - 20px)) {
        #t30 { background-color: green; }
    }

    @supports (background-color: rgb(0, 128, 0)) {
        #t31 { background-color: green; }
    }

    @supports (background: url("/blah")) {
        #t32 { background-color: green; }
    }

    @supports ((top: -webkit-calc(80% - 20px)) and (not (background-color: rgb(0, 128, 0)))) or (background: url("/blah")) {
        #t33 { background-color: green; }
    }

    @supports (background: invalid("/blah")) {
        #t34 { background-color: red; }
    }

    /* Nesting. */
    @supports (display: none) {
        @supports (display: deadbeef) {
            #t18 { background-color: red; }
        }
        @supports (display: inline) {
            #t19 { background-color: green; }
        }
        @supports (display: inline) {
        }
        @media all {
            #t20 { background-color: green; }
            @supports (display: inline) {
                #t21 { background-color: green; }
            }
        }
    }

    @media all {
        @supports (display: inline) {
            @media all {
                @supports (display: none) {
                    #t35 { background-color: green; }
                }
            }
        }
    }

    @media not all {
        @supports (display: none) {
            #t36 { background-color: red; }
        }
    }

    /* Invalid syntax error recovery */

    @supports (display: none);
    @supports (display: none) and ( (display: none) ) {
        #t37 { background-color: green; }
    }

    @supports (display: none)) ;
    @supports (display: none) {
        #t38 { background-color: green; }
    }

    @supports;
    @supports (display: none) {
        #t39 { background-color: green; }
    }

    @supports ;
    @supports (display: none) {
        #t40 { background-color: green; }
    }

    @supports (display: none)) {
        #t41 { background-color: red; }
    }
    @supports (display: none) {
        #t41 { background-color: green; }
    }

    @supports (display: )) {
        #t42 { background-color: red; }
    }
    @supports (display: none) {
        #t42 { background-color: green; }
    }

    @supports ((display: none) and {
        #t43 { background-color: red; }
    }
    @supports (display: none) {
        #t43 { background-color: red; }
    }
    );

    @supports ((display: none) and ;
    @supports (display: none) {
        #t44 { background-color: red; }
    }
    );

    @supports (#(display: none) {}
        #t45 { background-color: red; }
    );

    @supports ((display: none#) {}
        #t46 { background-color: red; }
    );

    @supports (#) or (display: none) {
        #t47 { background-color: green;; }
    }

</style>
</head>
<body>
<div id="test_container"></div>

<script>
    description("Test the @supports rule.");
    var numTests = 48;
    var untouchedTests = [1, 3, 5, 8, 12, 13, 14, 18, 22, 23, 24, 25, 28, 29, 34, 36, 43, 44, 45, 46]; // Tests whose content shouldn't change from the UNTOUCHED default.
    var UNTOUCHED = "rgb(0, 0, 255)";
    var APPLIED = "rgb(0, 128, 0)";

    var container = document.getElementById("test_container");
    for (var i=0; i < numTests; i++) {
        var div = document.createElement("div");
        div.id = "t" + i;
        div.className = "test";
        container.appendChild(div);
        shouldBeEqualToString("getComputedStyle(document.getElementById('t"+i+"')).backgroundColor", untouchedTests.indexOf(i) >= 0 ? UNTOUCHED : APPLIED);
    }

    test_container.parentNode.removeChild(test_container);
</script>
</body>
</html>
