third_party/WebKit/LayoutTests/css3/supports-expected.txt

Test the @supports rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.getElementById('t0')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t1')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t2')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t3')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t4')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t5')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t6')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t7')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t8')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t9')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t10')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t11')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t12')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t13')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t14')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t15')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t16')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t17')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t18')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t19')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t20')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t21')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t22')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t23')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t24')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t25')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t26')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t27')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t28')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t29')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t30')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t31')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t32')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t33')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t34')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t35')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t36')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t37')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t38')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t39')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t40')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t41')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t42')).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(document.getElementById('t43')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t44')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t45')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t46')).backgroundColor is "rgb(0, 0, 255)"
PASS getComputedStyle(document.getElementById('t47')).backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
